requirement.txt
Pillow==10.0.1
PyYAML==6.0.1
